I completed a first working prototype of jwt token authorization

+++ src/main/java/pl/web/Service/UsersService.java
@@ -1,9 +1,9 @@
 package pl.web.Service;
 
+import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import pl.web.Entity.Settings;
 import pl.web.Entity.User;
@@ -19,62 +19,66 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class UsersService {
+public class UsersService extends AccessChecking {
     private final UserRepository userRepository;
     private final SettingsRepository settingsRepository;
-    private final PasswordEncoder passwordEncoder;
 
     @Autowired
-    public UsersService(UserRepository userRepository, SettingsRepository settingsRepository, PasswordEncoder passwordEncoder) {
+    public UsersService(UserRepository userRepository, SettingsRepository settingsRepository) {
         this.userRepository = userRepository;
         this.settingsRepository = settingsRepository;
-        this.passwordEncoder = passwordEncoder;
     }
 
-    public ResponseEntity<?> removeUser(User user) {
+    public ResponseEntity<?> removeUser(@NotNull User user) {
         if (user.getEmail().isEmpty() || user.getUsername().isEmpty() || user.getId() == null) {
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-        } else {
-            Optional<User> userOptional = userRepository.findUserAllById(user.getId());
-            userRepository.delete(userOptional.get());
-            return ResponseEntity.ok().build();
         }
+        Optional<User> userOptional = userRepository.findUserAllById(user.getId());
+        if (notAccess(userOptional)) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        userRepository.delete(userOptional.get());
+        return ResponseEntity.ok().build();
     }
 
-    public ResponseEntity<UserResponse> getUserData(EmailModel emailModel) {
+    public ResponseEntity<?> getUserData(@NotNull EmailModel emailModel) {
         if (emailModel.getEmail().isEmpty()) {
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
         }
         Optional<User> userOptional = userRepository.findByEmail(emailModel.getEmail());
         return userOptional.map(user -> {
-            return ResponseEntity.ok(new UserResponse(
-                    userOptional.get().getUsername(),
-                    userOptional.get().getEmail(),
-                    userOptional.get().getId(),
-                    userOptional.get().getStatus(),
-                    null
-            ));
+            return ResponseEntity.ok(generateUserResponse(userOptional, true));
         }).orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
     }
 
-    public ResponseEntity<?> getUserData(IdModel idModel) {
-        if (idModel.getId() == null) {
-            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-        }
+    public ResponseEntity<?> getMyData(@NotNull IdModel idModel) {
         Optional<User> userOp = userRepository.findById(idModel.getId());
         if (userOp.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
         }
-        return ResponseEntity.ok(new UserResponse(
-                userOp.get().getUsername(),
-                userOp.get().getEmail(),
-                userOp.get().getId(),
-                userOp.get().getStatus()
-        ));
+        if (notAccess(userOp)) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        return ResponseEntity.ok(generateUserResponse(userOp, false));
+    }
+
+    public ResponseEntity<?> getUserData(@NotNull IdModel idModel) {
+        Optional<User> userOp = userRepository.findById(idModel.getId());
+        if (userOp.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        }
+        return ResponseEntity.ok(generateUserResponse(userOp, true));
+    }
+
+    private Object generateUserResponse(Optional<User> userOp, Boolean includeStatus) {
+        User user = userOp.orElseThrow();
+        Settings settings = settingsRepository.findByUserid(user.getId()).orElseThrow();
+
+        String userName = includeStatus && !settings.getShowName() ? null : user.getUsername();
+        String userEmail = includeStatus && !settings.getShowEmail() ? null : user.getEmail();
+        String userStatus = includeStatus && !settings.getShowStatus() ? null : user.getStatus();
+        Long userId = user.getId();
+
+        return new UserResponse(userName, userEmail, userId, userStatus);
     }
 
     public ResponseEntity<?> getAllUsers() {
-        System.err.println(passwordEncoder.encode("root"));
         List<User> users = userRepository.findAll();
         List<ListResponse> listResponses = new ArrayList<ListResponse>();
         if (users.isEmpty()) {
@@ -87,7 +91,7 @@ public class UsersService {
         return ResponseEntity.ok(listResponses);
     }
 
-    public ResponseEntity<?> getVisibilitySettings(IdModel idModel) {
+    public ResponseEntity<?> getVisibilitySettings(@NotNull IdModel idModel) {
         Optional<Settings> settingsOptional = settingsRepository.findByUserid(idModel.getId());
         if (settingsOptional.isEmpty()) {
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
